Split Autocomplete JS into modules
